Regres: Add workaround for Khronos Gitlab issues
It seems that once a change is no longer in the top N most recent changes, the server will start responding `server does not allow request for unadvertised object` when attempting to fetch the specific change by SHA.

To work around this, we need to fetch the branch by name, then checkout the specific change locally. As the Khronos seems to be particularly slow to respond today, I've limited the branch depth to 99, for now. This hopefully will not be a problem as we should typically only be updating to newer changes.

Also bumped the git command timeout to 15 mintutes as this repo can be _really_ slow.

Change-Id: Ie08165a6ca92e79d56bb4d2d5376410f2850a471