Negative baseVertex support
baseVertex is a signed integer that's allowed have a negative value.
In order to support it, the offset was moved from being applied
on the index buffer memory pointer directly to being applied
to the indices from the index buffer inside the Vertex routine.

Fixes all issues in the following subcategory using SwANGLE:
dEQP-GLES31.functional.draw_indirect.*

Bug: b/144353667
Change-Id: Ia70c258eab25c274553c4955dc2336a42d43f005