Clarify the link between subpixel precision and framebuffer size
The rasterization code expresses the render target coordinates using
a fixed point unsigned short. Currently, the integer portion uses 28
bits and the subpixel portion uses the remaining 4 bits. Increasing
subpixel accuracy would reduce the maximum render target size, as it
would reduce the number of bits available to represent the integer
portion of the render target coordinate in the fixed point number.
The framebuffer size limit we currently use is the minimum required
by the Vulkan spec, which is 4096, so this change:
- makes it clear that framebuffer size and subpixel precision are
  related to each other
- adds an assert if we try to increase framebuffer size beyond
  our internal limit of 65536, because of our outlines using
  unsigned shorts AND our internal limit of 8192, as long as
  we're still using the outline logic
- adds an assert if a large enough subpixel precision would make
  SwiftShader non spec compliant regarding the framebuffer size
- adds an assert if a low enough subpixel precision would make
  SwiftShader non spec compliant regarding subpixel precision

Until we change how we represent render target coordinates to
something other than a fixed point unsigned short, then the
subPixelPrecisionBits range is limited to [4, 9], 9 coming from:
(32 - (13 bits of outline precision)) / 2 == 19 / 2 == 9

Bug: b/141676114
Change-Id: Ib83d48e98dea8d011dd01debc5010ed7caa449d2