Android: Add an APK target for SwiftShader
This allows developers to create an APK in the Android tree
that just contains the SwiftShader native libraries. It can
then be used with upcoming platform changes that will allow
loading alternate SwiftShaders for debuggable apps.

With this change, running the following in AOSP:

  $ m SwiftShader

will generate an APK like this:

  $ jar tf out/target/product/flame/system/app/SwiftShader/SwiftShader.apk
  lib/arm64-v8a/vulkan.pastel.so
  lib/armeabi-v7a/vulkan.pastel.so
  resources.arsc
  AndroidManifest.xml
  META-INF/CERT.SF
  META-INF/CERT.RSA
  META-INF/MANIFEST.MF

Bug: b/143752295
Change-Id: Ia8b0ea52c50acfa25e580411b4b6ae61d788e209