Update patch set 1
Patch Set 1: Code-Review+2

I'm sympathetic to the cause, but this folder is synched to specific revisions of https://github.com/KhronosGroup/Vulkan-Headers/tree/master/include/vulkan to avoid any potential API/ABI breaking adjustments (which this file does not affect). If we do need a local diff, I'd expect it to be for something significant and there would be an expectation that it's temporary. Currently we have vulkan_fuchsia_extras.h, and it already mistakenly got removed during a sync.

It's even more likely that vulkan.hpp would find its way back into this folder on the next sync.

I do agree this file gets in the way and has no purpose for us though. I don't think it belongs in upstream Vulkan-Headers even. Yes, it's a Vulkan header, but it's not part of *the* headers defining the API. Ideally we'd push for its removal upstream.

Anyway, no objection to going ahead with this change as-is and looking into the longer-term solution if it keeps rising from the dead.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2