Fix vkQueuePresentKHR is pResults is not null.
pResults was not getting filled in correctly.

As per the spec:

pResults is an array of VkResult typed elements with
swapchainCount entries. Applications that do not
need per-swapchain results can use NULL for pResults.
If non-NULL, each entry in pResults will be set to the VkResult
for presenting the swapchain corresponding to the same index
in pSwapchains.

Change-Id: I42fcb6d96218d4dbfa60e857226c0aed72e7d8ca