[cmake] Add support for CMAKE_SYSTEM_NAME == "Fuchsia"
This patch modifies the CMakeLists.txt file and a few sources
in order to build SwiftShader with CMake with a fuchsia-specific
toolchain.

NOTE: This should not affect the build for other platforms!

This will make it easier to conditionally add Fuchsia-specific
sources to the build (see discussion on b/144687651 for details).

The BUILD.gn file will still be maintained in parallel for
Fuchsia-specific changes.

NOTE: As of now, the Fuchsia SDK doesn't provide a way for
      third-party code to access the framebuffer directly.

      Instead, client code should rely on Vulkan presentation
      surfaces or higher-level APIs provided by the compositor
      that are still in flux / unstable. Hence this will appear
      in a future CL.

- <EGL/eglplatform.h> is modified to define required types,
  even though there is no "official" display window type on
  Fuchsia at the moment.

- src/Main/FrameBufferFuchsia.cpp: An empty implementation of
  createFrameBuffer() that just returns NULL.

- third_party/SPIRV-Tools/: Small updates as well, matches the
  upstream pull request at:
    https://github.com/KhronosGroup/SPIRV-Tools/pull/3062

- third_party/llvm-7.0: Small update to avoid #warnings that
    are turned into compiler errors by default.

To test this:

  mkdir build-fuchsia && cd build-fuchsia
  cmake .. -DCMAKE_TOOLCHAIN_FILE=/path/to/FuchsiaSdkToolchain.cmake
  make -j32

Bug: 143122483 144687651
Change-Id: Ica3b98f5a8eb1370287e2bb4ff22ef938aaa4b49