Update patch set 1
Patch Set 1:

A bit more context: I was investigating b/144573625 and b/144692525, and I wanted to make sure that the optimization pass lists were the same as `spirv-opt -O` (they are).
It does highlight the fact that these may slip out of sync, and there are clearly inter-pass orderings that produce valid / or more performant spirv code.

Patch-set: 1