Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> I'm on the fence about this one. We like this list *now*, but there's no reason to believe that SPIRV-Tools' generic go-fast-mode is the ideal set for SwiftShader in general.

The counter to this argument is that there's already inter-pass dependencies (pass A needs to come before pass B), and it's a moderately safe bet that more of these will crop up over time.
It should be a given that RegisterPerformancePasses() generates a valid list of passes. Until we *need* to diverge, I feel it is better to stick with the 'official' set of optimization passes.

Happy to be convinced otherwise though!

Patch-set: 1