Make Routine::getEntry const as it's required by Coroutine
This is because Coroutines cache the acquired Routine, then multiple
threads will call getEntry on it, expecting it to return a pointer it
can jump to. Subzero's getEntry was lazily setting it's entry pointer in
getEntry, which is not thread-safe. This change makes getEntry const,
so we can assume it's thread-safe, and similar to LLVM, we cache the
entry pointer upon acquiring the routine.

Bug: b/130459196
Change-Id: Id9fa447ce55c7df6fb55e1c878ca021e825fecfa