CMakeLists: Add CMake rules for Vulkan debugger
`ENABLE_VULKAN_DEBUG` is the new option to enable building of the Vulkan debugger, which defaults to off.

`cppdap` is a new library that implements the debugger protocol.

Added a stub .cpp file to try and detect any build systems that are recursively globbing into the new src/Vulkan/Debug directory.

Actual debugger implementation will come in future changes.

Bug: b/145351270
Change-Id: I48619f1d8352c0525c8adbaa4bc3f2373a237d61