Vulkan: Add debugging of the command buffer.
Add a debugger context and server to vk::Device, along with a getter for acquiring the debugger context.

Generate a synthetic file containing all the Vulkan commands in the command buffer, and allow the debugger to single line step over these.

All of this is no-op unless ENABLE_VK_DEBUGGER is defined at compile time, and the VK_DEBUGGER_PORT env var is set at run time.

Bug: b/145351270
Change-Id: I8bea398a6c08d4cb23d76172dbca2a08ae109a6d