Regres: Bump the deqp basic fence test timeout
I suspect this is the cause of the test flakes.
We "only" had 10ms to submit the queue that signals the fence. I suspect that when the regres machine is under strain, system stalls could exceed this timeout.

Bumped the `LONG_FENCE_WAIT` to 1s.

Note: Assuming this fixes the flakes, this is should be upstreamed to dEQP. We may want to land this while its being reviewed.
Change-Id: I76278fded7bb9fc93669896f9a04dd65e835aaa8