Subzero: fix alloca return address allocation
Only breaks in Om1 optimization mode because this code path is specific
to Om1. Broke after 01386d175 was merged because maxOutArgsSizeBytes()
now returns 32 bytes for the shadow store, so this is a rarely exercised
code path. Problem is that on x64, the new register being created needs
to match the size of the destination register, so 64 bits, not 32 as it
was asssuming.

Bug: b/145685985
Change-Id: I298edd5b4a286a242b351cc6267d70ecb37943fb