Update patch set 1
Patch Set 1: Code-Review-1

Absolutely valuable as an experiment, so thanks for this, but my expectation is that it will help reinforce that we don't want to go this route. Anyway, I'm keeping an open mind. And some users may prefer a performance hit over the memory usage of full decompression (but we have other promising options for that).

Let's get back to this when we have something we can use as a benchmark...

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1