Update patch set 8
Patch Set 8:

> Patch Set 7:
> I strongly back this statement. I don't want my brain to be muddled with implementation details...

In the CL that introduced dbg::ID, you argued for the use of a technically redundant 'inline' to indicate that you really desire the class to behave as efficiently as 'int', both to the compiler and developers. I fully support that argument. I think a class like this is an exception to the rule that interface and implementation should be separated, because they're inevitably tightly intertwined if performance is a concern.

In fact when there's a getter method implementation that's not in-class defined, my brain immediately starts to wonder if it does something non-trivial. I have to go and look it up anyway to be sure.

So we should be conscious about both these kinds of brain "muddlings" and weigh them against each other, instead of attempting to apply a one-size-fits-all rule. I don't think you were lazy at all writing that CL, but implicitly applying that tradeoff.

> > Explicitly marking the classes 'final' also seems useful to me if we're revisiting this anyway.
> 
> Humm, maybe. Again, feels orthogonal to inlining (at least to me).

Definitely! Sorry about conflating it. I'm just arguing that if we have a class where it's not an implementation detail that we desire everything to be inlined, but more of a design goal, we probably also don't want to allow inheritance which could break that. It's just another way to make developers aware of the class's purpose and intended usage.

> The SwiftShader codebase currently has exactly 0 final classes. Perhaps if we are to start adding these, we should look more broadly across the codebase for candidates classes?

Sure, but note that use of 'final' is very scarce in most code bases, as far as I'm aware.

I don't think it's a coincidence that https://stackoverflow.com/questions/16808184/use-cases-for-final-classes mentions an int-like class as an example use case of it. The only one, in fact.

> FWIW - I'd personally only use 'finally' if there's some complex behavior that you *really* don't want people to derive from. There's a few cases around preventing functions from being overridden and the like, but I'd say these cases are rare.
> However, if the codebase's policy is push strongly for composition instead of inheritance, perhaps the default should be to mark everything as 'final' except those few classes we know we need to inherit from. I know of a few places we use inheritance that certainly feels a bit icky. Perhaps this would help us re-evaluate and justify these places where inheritance is used.

While it's definitely a good idea to re-evaluate the choice between composition and inheritance in various places, I'd be extremely careful about having any of that result in the use of 'final'. Making a class final is an indication that if one wants a similar class but they'd have to inherit from it, they're doing it wrong and must create a new class instead. That's not the signal we want to give when it's just a change in design choice. No matter how well argumented, if it uses inheritance today it can't be wrong to that kind of degree to prohibit it now and forever. While one could argue that 'final' can be removed again, nobody will remember in several years that using inheritance before didn't cause disaster, but it will feel like something set in stone that shouldn't even cross their mind to change it.

Patch-set: 8
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>