Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> (2 comments)
> 
> > Patch Set 10:
> > 
> > (4 comments)
> > 
> > > Patch Set 10:
> > > 
> > > (6 comments)
> > > 
> > > Getting pretty close!
> > > 
> > > Note that applying clang-format a second time creates additional changes.
> > 
> > This is troubling (along with '>>' turning into '> >').
> > What do you see when you run `clang-format --version`?
> 
> google3-trunk

As I mentioned in 39428, I think we're going to have to move to a fixed public release of clang-format.
google3-trunk will not be available to kokoro, and apparently wants to be different (we had issues with this for GAPID too).
I can try and make that less painful by adding scripts to fetch the particular revision into a tools directory for the given OS.

Patch-set: 11