Prefer alignas() over ALIGN()
Since the alignas() specifier is part of the type, it cannot be used in
a typedef, and compiler-specific attributes have to be used. We
abstracted these into the ALIGN() macro before, but used them in other
cases where alignas() can be used.

Using it where possible avoids formatting weirdness with clang-format.

Bug: b/144825072
Change-Id: If4185df296f0b525dd811103ac5bf1191833b154