Exclude padding bytes from image size limit
Surface::size() adds 4 padding bytes due to reading byte4 as 8 bytes for
efficient unpacking to short4 (less efficient prior to SSE 4.1 pmovzxbw
instruction). This padding is never addressed separately so the offset
can't overflow a signed 32-bit integer if we keep the total size minus
the padding below 2 GiB.

On the OpenGL side we currently impose a further limit of 1 GiB, but
this should also take the padding into account to allow for effective
image data up to this limit (e.g.  8192 * 8192 of four samples of 32-bit
elements). Note that our scanline limit is also 8192 so allowing more
than 1 GiB would run into that limit instead.

Bug: chromium:835299
Bug: b/145229887
Change-Id: I4280fca88584235c0da6282ca622ee15b31bc34b