WIP: add Vulkan gn targets for subzero and llvm
Adds the following targets:
* swiftshader_llvm_libvulkan
* swiftshader_subzero_libvulkan
* swiftshader_llvm_icd_file
* swiftshader_subzero_icd_file

and aliases existing targets:
* swiftshader_libvulkan to swiftshader_llvm_libvulkan
* icd_file to swiftshader_llvm_icd_file

This would work now, but is probably not what we really want. The thing
is that for the LLVM build, it outputs vk_swiftshader.dll, rather than
vk_swiftshader_llvm.dll. Rather than do it this way, let's first fix
Chrome code that hard-codes the DLL name, and make sure it uses the name
in the ICD file instead. Then, we are free to modify these target names
as we wish, so we can make the LLVM output vk_swiftshader_llvm.dll.