VulkanUnitTests: Fix SPIR-V validation issue
`LoopDivergentMergePhi` was attempting to use the same merge block for a `OpLoopMerge` and a `OpSelectionMerge`. This is not legal, and was caught by improvements to the SPIR-V tools validator.

The fix is simple: Add a new merge block for the inner `OpSelectionMerge` that immediately jumps to the merge block for the `OpLoopMerge`.

Bug: b/123642959
Bug: b/133440380
Bug: b/133481698
Change-Id: I8b5da04a1cd2db3e44bb5b50f1619a8565e5bc9d