Update patch set 26
Patch Set 26: Code-Review+2

> Patch Set 26:
> 
> > Patch Set 26: Code-Review-1
> > 
> > (1 comment)
> > 
> > Sorry about the lengthy discussion (FWIW I found it very interesting to cover all pros/cons), but unless there are new design requirements I've made up my mind about this
> 
> Let's discuss this in our 1:1 tomorrow.
> 
> Putting aside all personal preference arguments, I'm almost certain that the GAPID use case *requires* a separate debugger instance per device. While I don't quite follow how globals solve the inter-process CommandBuffer problem you've described, I'm willing to drop CommandBuffer debugging if this is a blocker for making progress here.
> 
> Also: I'm not sure if you've checked the latest patch sets, but I also want to point out that there is a separate debugger context per device now. There is no longer a debugger singleton.

Thanks for the offline chat. I'm going to approve this, but for the records I'll list my final reservations:

Testing GAPID and Radial's tools is great value, and the main reason why I'm willing to give you what you think is needed. Perhaps just having the SpirvShader class take the debugger parameter suffices for Radial's unit testing needs though. And for GAPID I'm not sure I fully agree with the overhead argument against multi-process testing instead of multi-device. Replaying any non-empty trace is going to dominate execution time, not the process start time.

The argument that AMD's hardware can likely support multi-device debugging is definitely interesting, but is not a guarantee that they'll actually implement that in the driver. Game developers most certainly only have a need for a single debugger per process, and so any amount of complication from supporting debugging for multiple ones is unwarranted. As for mobile devices, while some don't support shader debugging at all, those on the verge of providing support for it might be scared off by an extension which requires multi-device debugging support. We might end up with a second extension, which leads to fragmentation or people just using the single-debugger one which suits their needs fine.

I'm not enamoured by the idea that this could be a SwiftShader-only feature. There's great value in having drop-in replacement compatibility between implementations. We'd have to pay the burden of supporting this feature indefinitely. While it's hopefully a small one for this feature, I'm very wary about the precedent this sets. It's a slippery slope that can lead to other invasive changes that add more complexity.

Which brings me back to my other point. Conceptually a debugger should have no impact on the code that's being debugged, only on the compiler. In our case SwiftShader is both the code being debugged and the shader compiler (and the command buffer interpreter), but those should be kept separate in my view. So if the design requirements only mandate support for one debugger per process, as usual, it is totally reasonable to provide it to the parts of the code base that need it by singleton.

Anywho, let's see how this plays out. I'll re-evaluate this when we do the process separation.

Patch-set: 26
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2