Add the PowerVR SDK as a git submodule
The primary purpose of doing this is to have all the PowerVR SDK
examples in one place and remain up-to-date. This change also lets us
simplify a bit of our CMakeLists.txt as we only need to use
add_subdirectory to build everything we need.

The PowerVR SDK examples are built by explicitly enabling
SWS_BUILD_PVR_EXAMPLES (either through a cmake gui, or
-DSWS_BUILD_PVR_EXAMPLES=ON on the command line). You must also specify
a window system for the examples to use with something like
"-DPVR_WINDOW_SYSTEM=X11".

Several cmake option flags were renamed to start with "SWS" to avoid
name collisions with other cmake files in subdirectories we cannot
control.

Bug: b/141361933
Change-Id: I6d2fac579f62ff4bc211d5fd5127fb670eb9b4bf