A2R10G10B10 Support
A2B10G10R10 (RGBA) was already supported by
SwiftShader, but A2R10G10B10 (BGRA) was not. Most of
this cl is trivial, as it simply adds an equivalent
version of the new formats from the code used for the
already supported formats, with the R and B channels
swapped.

The only new piece of code is rounding for 1010102
formats at the top of the PixelRoutine::writeColor()
function. There was already rounding for 8 bit formats,
but not 1010102, which led to potential off by 1 errors
in the output, which is fairly large when it happens
on the 2 bit alpha channel. This fixes one of the
dEQP-VK.pipeline.blend.*a2r10g10b10* tests.

Tests: dEQP-VK.*a2r10g10b10*
Bug: b/142661203

Change-Id: Ifcae17aecafab3ea7967fdc755391ddd5e651ca5