Allow verbal bug IDs
In rare cases the requirement to have a Buganizer or Monorail bug ID
in the commit message can be unreasonable. For example when the build
is broken and the fix is trivial so the creation of a tracker bug would
needlessly take valuable time. Another example would be fixing a
spelling mistake in a comment and we don't want to spam the tracker
thread with a commit notification.

For such cases, accept bug info lines containing no digits.

Bug: b/141892461