Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Your regex currently permits: 'Bug: ' (trailing whitespace)
> > That could be fixed with:
> >  "^(Bug|Issue|Fixes):(\s?)(((b\/)|(\w+:))([0-9]+)|[\w]+[\w\s]*)$|(^Regres:)"
> > 
> > Or, we could just have an explicit "Bug: none":
> >  "^(Bug|Issue|Fixes):(\s?)(((b\/)|(\w+:))([0-9]+)|[Nn]one)$|(^Regres:)"
> > And expect the reason to be contained in the proper subject / message.
> 
> It technically permits a lot of cases that could be considered abuse, but it's not intended to guard against all of that. If the bug ID format intentionally doesn't match that of Buganizer or Monorail, the author and reviewer still have a responsibility to make it reasonably useful (like the rest of the commit message).
> 
> I'm simply making the assumption that if there's a number, it must be in Buganizer or Monorail format, and catch common mistakes, e.g. a number without prefix, or multiple IDs on the same line. Your version doesn't prevent those, and doesn't match e.g. "Bug: fix build!".

Sure. I was simply trying to point out that the rules are growing increasingly complex over time, and perhaps instead of encouraging a weird psuedo-description-on-bug-field-that-isn't-a-bug-id-which-fails-if-it-contains-a-number, we enforce a real bug id, or a allow a simple no-bug token and expect the rest of the CL message to clearly describe why. At least this is simple to document.

I don't particularly care either way, and I'm certainly not going to hold up this change.

Patch-set: 1
Reviewer: Gerrit User 7183 <7183@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>