Replace positive/negative_inf() with infinity()
This matches the std::numeric_limits method name.

Negative infinity can be obtained using -infinity(). This comes at a
minor JIT run-time cost, and the backend may or may not perform constant
folding. This can be optimized by having Reactor perform the folding
(which would still be cheaper than having a capable backend do it).

Also replace the use of C99 INFINITY with the C++ constant.

Bug: b/140302841
Change-Id: Ifd5e5d1fa4fa2d9ccce6d334af22903ea7597a69