Refactor binary group operations
Instead of deducing the binary operation type from the identity value,
which used a type alias, just specify it explicitly. This enables
passing the initialization value used as the identity, instead of
necessarily an identity value vector itself, as well as omitting the
type alias. In turn this allowed omitting the scope brackets, resulting
in a significant code compaction without compromising readability.

Bug: b/142002682
Change-Id: I23d6d984bbfdf47af72108adbc942992007fc3c0