Vulkan: Disable SPIR-V opts & val when debugging
Optimization passes are likely to damage debug information, and reorder instructions. We will likely never want to optimize shaders inside the driver when debugging.

Validation currently has a number of bugs which need to be fixed before debug information can be used. For now, just disable shader validation when we are debugging.

Bug: b/145351270
Change-Id: I71827c2bb7cf34ecfd7e55b0ca8b3189e460b150