Vulkan: Disable SPIR-V preprocessing when debugging
Optimization passes are likely to damage debug information, and reorder instructions. We will likely never want to optimize shaders inside the driver when debugging.

SPIR-V tools currently has a number of issues consuming shaders that use the OpenCL.Debug.100 extension instructions which need to be fixed before these instructions can be passed to it. For now, just disable shader pre-processing when we are debugging.

Bug: b/145351270
Change-Id: I71827c2bb7cf34ecfd7e55b0ca8b3189e460b150