.vscode: Use compile_commands.json for build env
Instead of listing a hard-coded set of defines and include paths  for vscode, use the `CMAKE_EXPORT_COMPILE_COMMANDS` flag to generate a `compile_commands.json` file, which lists the build flags for each cpp file.

This reduces maintainance, and keeps the IDE in sync with the current build configuration.

Bug: none
Change-Id: I2161aee61705d11029d8a5a2127130628f2392f6