Replace all UNIMPLEMENTED() with UNSUPPORTED()
Most cases of UNIMPLEMENTED() are actually for functionality that we do
not support, and correctly advertise as such. It's expected to be an
application bug if these asserts are hit, and not an indication of
something we must still implement to be a compliant Vulkan
implementation.

This change will be immediately followed up by one which restores
UNIMPLEMENTED() where it was used correctly. This two-change approach
makes it much easier to see what still needs to get done.

Bug: b/131243109
Change-Id: Id09b5de5bc73fabfdc3199dcd3091cb72283be65